backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
supabase==2.10.0
python-dotenv==1.0.1
openai==1.58.1
requests==2.32.3
twilio==9.3.7
python-multipart==0.0.20
